massive refactoring of the crawler to make it into a separate object from the app.js.
